backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7
pydantic>=2.3.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
redis>=5.0.0
meilisearch>=0.25.0
boto3>=1.28.40
python-dotenv>=1.0.0